move to redux folder and fix import
