+++ frontend/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { Provider } from 'react-redux'
-import store from './store'
+import store from './redux/store'
 
 import './bootstrap.min.css'
 import './index.css'
